Shaded gcp and oss bundle files.
